abstract away common intro for proposal.js
