+++ tests/templates/proposal.template.js
@@ -1,40 +1,3 @@
-console.log("unlocking accounts of token holders");
-personal.unlockAccount(eth.accounts[0], "Write here a good, randomly generated, passphrase!");
-personal.unlockAccount(eth.accounts[1], "Write here a good, randomly generated, passphrase!");
-personal.unlockAccount(eth.accounts[2], "Write here a good, randomly generated, passphrase!");
-personal.unlockAccount(eth.accounts[3], "Write here a good, randomly generated, passphrase!");
-personal.unlockAccount(eth.accounts[4], "Write here a good, randomly generated, passphrase!");
-personal.unlockAccount(eth.accounts[5], "Write here a good, randomly generated, passphrase!");
-personal.unlockAccount(eth.accounts[6], "Write here a good, randomly generated, passphrase!");
-// set coinbase to something other than service provider and proposal creator
-web3.miner.setEtherbase(eth.accounts[5]);
-
-var serviceProvider = eth.accounts[0];
-var proposalCreator = eth.accounts[1];
-var testMap = {};
-
-function checkWork() {
-    miner.start(1);
-    // console.log("Creator Balance: web3.fromWei(eth.getBalance(proposalCreator)));
-    admin.sleepBlocks(3);
-    miner.stop(0);
-}
-
-function bigDiff(astr, bstr) {
-    return Math.round((new BigNumber(astr)).minus(new BigNumber(bstr)));
-}
-
-function addToTest(name, value) {
-    testMap[name] = value;
-    console.log("'" + name + "' = " + value);
-}
-
-function testResults() {
-    console.log("Test Results: " + JSON.stringify(testMap));
-}
-
-
-
 var dao = web3.eth.contract($dao_abi).at('$dao_address');
 var offer = web3.eth.contract($offer_abi).at('$offer_address');
